Moved and refactored image source check
Moved the image source check from the Editor-page to the Editor-component, and refactored to use component-variables instead. This seems to fix a rare bug where sometimes, if a user reloaded very fast after initially loading in successfully, the image would stay but fail to upload, and the user would not be redirected to the home page
